Updated for 1.3

+++ Source/Gradual Romance/AttractionCalculator_RelationshipLimit.cs
@@ -13,7 +13,7 @@ namespace Gradual_Romance
 
         public override float Calculate(Pawn observer, Pawn assessed)
         {
-            return GradualRomanceMod.numberOfRelationships /
+            return (float) GradualRomanceMod.numberOfRelationships /
                    (RelationshipUtility.GetAllPawnsRomanticWith(observer).Count + 1);
         }
     }
